# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trust Bank Marketing & Sales Suite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns (optionally filter by channel/status)
      parameters:
        - in: query
          name: channel
          schema: { type: string, enum: [digital, sms, email, branch] }
        - in: query
          name: status
          schema: { type: string, enum: [active, paused, ended] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /interactions:
    get:
      operationId: listInteractions
      tags: [interactions]
      parameters:
        - in: query
          name: leadId
          schema: { type: integer }
        - in: query
          name: salesOfficerId
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interaction"
    post:
      operationId: createInteraction
      tags: [interactions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInteractionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interaction"
...
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      parameters:
        - in: query
          name: stage
          schema: { type: string, enum: [prospect, negotiation, won, lost] }
        - in: query
          name: salesOfficerId
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [deals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDealBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}:
    patch:
      operationId: updateDeal
      tags: [deals]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDealBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
...
components:
  schemas:
...
    Campaign:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        channel: { type: string, enum: [digital, sms, email, branch] }
        startDate: { type: string }
        endDate: { type: string, nullable: true }
        budget: { type: number }
        spent: { type: number }
        status: { type: string, enum: [active, paused, ended] }
        targetAudience: { type: string }
        ownerId: { type: integer }
        ownerName: { type: string }
        leadsGenerated: { type: integer }
        conversions: { type: integer }
      required: [id, name, channel, startDate, budget, spent, status, targetAudience, ownerId, ownerName, leadsGenerated, conversions]
    CreateCampaignBody:
      type: object
      properties:
        name: { type: string }
        channel: { type: string, enum: [digital, sms, email, branch] }
        startDate: { type: string }
        endDate: { type: string, nullable: true }
        budget: { type: number }
        targetAudience: { type: string }
        status: { type: string, enum: [active, paused, ended] }
      required: [name, channel, startDate, budget, targetAudience, status]
    UpdateCampaignBody:
      type: object
      properties:
        name: { type: string }
        channel: { type: string, enum: [digital, sms, email, branch] }
        endDate: { type: string, nullable: true }
        budget: { type: number }
        spent: { type: number }
        status: { type: string, enum: [active, paused, ended] }
        targetAudience: { type: string }
...
    Interaction:
      type: object
      properties:
        id: { type: integer }
        leadId: { type: integer }
        leadName: { type: string }
        type: { type: string, enum: [call, meeting, branch_visit, sms, email] }
        notes: { type: string }
        salesOfficerId: { type: integer }
        salesOfficerName: { type: string }
        nextFollowUp: { type: string, nullable: true }
        createdAt: { type: string }
      required: [id, leadId, leadName, type, notes, salesOfficerId, salesOfficerName, createdAt]
    CreateInteractionBody:
      type: object
      properties:
        leadId: { type: integer }
        type: { type: string, enum: [call, meeting, branch_visit, sms, email] }
        notes: { type: string }
        nextFollowUp: { type: string, nullable: true }
      required: [leadId, type, notes]
...
    Deal:
      type: object
      properties:
        id: { type: integer }
        customerName: { type: string }
        customerPhone: { type: string }
        productType: { type: string }
        value: { type: number }
        stage: { type: string, enum: [prospect, negotiation, won, lost] }
        expectedCloseDate: { type: string }
        salesOfficerId: { type: integer }
        salesOfficerName: { type: string }
        leadId: { type: integer, nullable: true }
        notes: { type: string, nullable: true }
        createdAt: { type: string }
      required: [id, customerName, customerPhone, productType, value, stage, expectedCloseDate, salesOfficerId, salesOfficerName, createdAt]
    CreateDealBody:
      type: object
      properties:
        customerName: { type: string }
        customerPhone: { type: string }
        productType: { type: string }
        value: { type: number }
        stage: { type: string, enum: [prospect, negotiation, won, lost] }
        expectedCloseDate: { type: string }
        leadId: { type: integer, nullable: true }
        notes: { type: string, nullable: true }
      required: [customerName, customerPhone, productType, value, stage, expectedCloseDate]
    UpdateDealBody:
      type: object
      properties:
        stage: { type: string, enum: [prospect, negotiation, won, lost] }
        value: { type: number }
        expectedCloseDate: { type: string }
        notes: { type: string, nullable: true }
...