# lib/api-zod/src/generated/types/updateDealBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trust Bank Marketing & Sales Suite API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateDealBodyStage } from "./updateDealBodyStage";

export interface UpdateDealBody {
  stage?: UpdateDealBodyStage;
  value?: number;
  expectedCloseDate?: string;
  notes?: string | null;
}